refactor: make spec created more reasonable
Old spec related code is splited messed, lots of fields in spec
have relationship with each other, simple make them as general
funtion is wrong, I try to make spec files splited from runtime
spec struct, make it more reasonable and remove redundancy code.
