+++ daemon/mgr/spec_annotations.go
@@ -4,16 +4,16 @@ import (
 	"context"
 	"strconv"
 
+	specs "github.com/opencontainers/runtime-spec/specs-go"
 	"github.com/sirupsen/logrus"
 )
 
 // setupAnnotations extracts other related options from HostConfig and locate them in spec's annotations which will be dealt by vendored runc.
-func setupAnnotations(ctx context.Context, meta *ContainerMeta, spec *SpecWrapper) error {
-	s := spec.s
-
-	r := meta.HostConfig.Resources
-
-	s.Annotations = make(map[string]string)
+func setupAnnotations(ctx context.Context, c *ContainerMeta, s *specs.Spec) error {
+	if s.Annotations == nil {
+		s.Annotations = make(map[string]string)
+	}
+	r := c.HostConfig.Resources
 
 	if r.MemoryWmarkRatio != nil {
 		s.Annotations["__memory_wmark_ratio"] = strconv.FormatInt(*r.MemoryWmarkRatio, 10)
@@ -28,7 +28,7 @@ func setupAnnotations(ctx context.Context, meta *ContainerMeta, spec *SpecWrappe
 	s.Annotations["__schedule_latency_switch"] = strconv.FormatInt(r.ScheLatSwitch, 10)
 
 	// add additional spec annotations
-	annotations := meta.Config.SpecAnnotation
+	annotations := c.Config.SpecAnnotation
 	for k, v := range annotations {
 		if _, exist := s.Annotations[k]; exist {
 			logrus.Warnf("Duplicate spec annotation: %s=%s", k, v)
