+++ daemon/mgr/spec_hook.go
@@ -2,53 +2,72 @@ package mgr
 
 import (
 	"context"
+	"os"
+	"path/filepath"
 	"sort"
+	"strconv"
 	"strings"
 
+	"github.com/alibaba/pouch/storage/quota"
 	specs "github.com/opencontainers/runtime-spec/specs-go"
 )
 
 //setup hooks specified by user via plugins, if set rich mode and init-script exists set init-script
-func setupHook(ctx context.Context, c *ContainerMeta, spec *SpecWrapper) error {
-	// if no init script specified and no hook plugin setup, skip it
-	if (!c.Config.Rich || c.Config.InitScript == "") && len(spec.argsArr) == 0 {
-		return nil
+func setupHook(ctx context.Context, c *ContainerMeta, specWrapper *SpecWrapper) error {
+	s := specWrapper.s
+	if s.Hooks == nil {
+		s.Hooks = &specs.Hooks{
+			Prestart:  []specs.Hook{},
+			Poststart: []specs.Hook{},
+			Poststop:  []specs.Hook{},
+		}
 	}
 
-	if spec.s.Hooks == nil {
-		spec.s.Hooks = &specs.Hooks{}
-	}
-
-	if spec.s.Hooks.Prestart == nil {
-		spec.s.Hooks.Prestart = []specs.Hook{}
-	}
-
-	if len(spec.argsArr) > 0 {
+	// setup plugin hook, if no hook plugin setup, skip this part.
+	argsArr := specWrapper.argsArr
+	prioArr := specWrapper.prioArr
+	if len(argsArr) > 0 {
 		var hookArr []*wrapperEmbedPrestart
-		for i, hook := range spec.s.Hooks.Prestart {
+		for i, hook := range s.Hooks.Prestart {
 			hookArr = append(hookArr, &wrapperEmbedPrestart{-i, append([]string{hook.Path}, hook.Args...)})
 		}
-		priorityArr := spec.prioArr
-		argsArr := spec.argsArr
-		for i, p := range priorityArr {
+		for i, p := range prioArr {
 			hookArr = append(hookArr, &wrapperEmbedPrestart{p, argsArr[i]})
 		}
 		sortedArr := hookArray(hookArr)
 		sort.Sort(sortedArr)
-		spec.s.Hooks.Prestart = sortedArr.toOciPrestartHook()
+		s.Hooks.Prestart = append(s.Hooks.Prestart, sortedArr.toOciPrestartHook()...)
 	}
 
-	args := strings.Fields(c.Config.InitScript)
-	if len(args) == 0 {
-		return nil
+	// setup rich mode container hoopk, if no init script specified and no hook plugin setup, skip this part.
+	if c.Config.Rich && c.Config.InitScript != "" {
+		args := strings.Fields(c.Config.InitScript)
+		if len(args) > 0 {
+			s.Hooks.Prestart = append(s.Hooks.Prestart, specs.Hook{
+				Path: args[0],
+				Args: args[1:],
+			})
+		}
 	}
 
-	preStartHook := specs.Hook{
-		Path: args[0],
-		Args: args[1:],
-	}
+	// setup diskquota hook, if rootFSQuota not set skip this part.
+	rootFSQuota := quota.GetDefaultQuota(c.Config.DiskQuota)
+	if rootFSQuota != "" {
+		qid := "0"
+		if c.Config.QuotaID != "" {
+			qid = c.Config.QuotaID
+		}
 
-	spec.s.Hooks.Prestart = append(spec.s.Hooks.Prestart, preStartHook)
+		target, err := os.Readlink(filepath.Join("/proc", strconv.Itoa(os.Getpid()), "exe"))
+		if err != nil {
+			return err
+		}
+
+		s.Hooks.Prestart = append(s.Hooks.Prestart, specs.Hook{
+			Path: target,
+			Args: []string{"set-diskquota", c.BaseFS, rootFSQuota, qid},
+		})
+	}
 
 	return nil
 }
